Refactor file conflict tests
