+++ spec/integration/hanami/cli/commands/app/generate/part_spec.rb
@@ -5,13 +5,10 @@ require "hanami"
 RSpec.describe "Hanami generate part integration", :app do
   let(:fs) { Hanami::CLI::Files.new(memory: false, out: out) }
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
 
   subject(:command) do
-    Hanami::CLI::Commands::App::Generate::Part.new(fs: fs, out: out, err: err)
+    Hanami::CLI::Commands::App::Generate::Part.new(fs: fs, out: out)
   end
-
-  def error_output = err.string.chomp
 
   around do |example|
     Dir.mktmpdir do |dir|
@@ -63,22 +60,6 @@ RSpec.describe "Hanami generate part integration", :app do
       }.not_to raise_error
 
       expect(fs.exist?("app/views/parts/book.rb")).to be(true)
-    end
-  end
-
-  context "error handling" do
-    let(:file_path) { "app/views/parts/existing.rb" }
-
-    it "handles file conflicts" do
-      fs.mkdir("app/views/parts")
-      fs.write(file_path, "# existing content")
-
-      expect do
-        command.call(name: "existing")
-      end.to raise_error SystemExit do |exception|
-        expect(exception.status).to eq 1
-        expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-      end
     end
   end
 end
