+++ spec/unit/hanami/cli/commands/app/generate/action_spec.rb
@@ -3,10 +3,9 @@
 require "hanami"
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Action, :app do
-  subject { described_class.new(fs: fs, out: out, err: err) }
+  subject { described_class.new(fs: fs, out: out) }
 
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:inflector) { Dry::Inflector.new }
   let(:app) { Hanami.app.namespace }
@@ -18,8 +17,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Action, :app do
   def output
     out.rewind && out.read.chomp
   end
-
-  def error_output = err.string.chomp
 
   shared_context "with existing files" do
     before do
@@ -36,65 +33,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Action, :app do
           expect(output).to include("Created app/actions/#{controller}/#{action}.rb")
           expect(output).to include("Created app/views/#{controller}/#{action}.rb")
           expect(output).to include("Created app/templates/#{controller}/#{action}.html.erb")
-        end
-      end
-    end
-
-    context "with existing action file" do
-      let(:file_path) { "app/actions/#{controller}/#{action}.rb" }
-
-      before do
-        within_application_directory do
-          fs.write(file_path, "")
-        end
-      end
-
-      it "exits with error message" do
-        expect do
-          within_application_directory { generate_action }
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
-    end
-
-    context "with existing view file" do
-      let(:file_path) { "app/views/#{controller}/#{action}.rb" }
-
-      before do
-        within_application_directory do
-          fs.write(file_path, "")
-        end
-      end
-
-      it "exits with error message" do
-        expect do
-          within_application_directory { generate_action }
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
-    end
-
-    context "with existing template file" do
-      let(:file_path) { "app/templates/#{controller}/#{action}.html.erb" }
-
-      before do
-        within_application_directory do
-          fs.write(file_path, "")
-        end
-      end
-
-      it "exits with error message" do
-        expect do
-          within_application_directory do
-            generate_action
-          end
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
         end
       end
     end
