+++ spec/unit/hanami/cli/commands/app/generate/migration_spec.rb
@@ -3,17 +3,14 @@
 require "hanami"
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Migration, :app do
-  subject { described_class.new(fs: fs, err: err) }
+  subject { described_class.new(fs: fs) }
 
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
 
   def output
     out.string.strip
   end
-
-  def error_output = err.string.chomp
 
   let(:app) { Hanami.app.namespace }
 
@@ -65,23 +62,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Migration, :app do
                          "Name must contain only letters, numbers, and underscores."
                        ))
     end
-
-    context "with existing file" do
-      let(:file_path) { "config/db/migrate/20240713140600_create_posts.rb" }
-
-      before do
-        fs.write(file_path, "existing content")
-      end
-
-      it "exits with error message" do
-        expect do
-          subject.call(name: "create_posts")
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
-    end
   end
 
   context "generating for a slice" do
@@ -93,25 +73,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Migration, :app do
 
       expect(fs.read("slices/main/config/db/migrate/20240713140600_create_posts.rb")).to eq migration_file_contents
       expect(output).to eq("Created slices/main/config/db/migrate/20240713140600_create_posts.rb")
-    end
-
-    context "with existing file" do
-      let(:file_path) { "slices/main/config/db/migrate/20240713140600_create_posts.rb" }
-
-      context "with existing file" do
-        before do
-          fs.write(file_path, "existing content")
-        end
-
-        it "exits with error message" do
-          expect do
-            subject.call(name: "create_posts", slice: "main")
-          end.to raise_error SystemExit do |exception|
-            expect(exception.status).to eq 1
-            expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-          end
-        end
-      end
     end
   end
 end
