+++ spec/unit/hanami/cli/commands/app/generate/operation_spec.rb
@@ -1,18 +1,15 @@
 # frozen_string_literal: true
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Operation, :app do
-  subject { described_class.new(fs: fs, out: out, err: err) }
+  subject { described_class.new(fs: fs, out: out) }
 
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:app) { Hanami.app.namespace }
 
   def output
     out.string.chomp
   end
-
-  def error_output = err.string.chomp
 
   context "generating for app" do
     it "generates an operation without a namespace, with a recommendation" do
@@ -80,23 +77,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Operation, :app do
       expect(fs.read("app/admin/books/add.rb")).to eq(operation_file)
       expect(output).to include("Created app/admin/books/add.rb")
     end
-
-    context "with existing file" do
-      let(:file_path) { "app/admin/books/add.rb" }
-
-      before do
-        fs.write(file_path, "existing content")
-      end
-
-      it "exits with error message" do
-        expect do
-          subject.call(name: "admin.books.add")
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
-    end
   end
 
   context "generating for a slice" do
@@ -144,24 +124,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Operation, :app do
 
       expect(fs.read("slices/main/admin/books/add.rb")).to eq(operation_file)
       expect(output).to include("Created slices/main/admin/books/add.rb")
-    end
-
-    context "with existing file" do
-      let(:file_path) { "slices/main/admin/books/add.rb" }
-
-      before do
-        fs.mkdir("slices/main")
-        fs.write(file_path, "existing content")
-      end
-
-      it "exits with error message" do
-        expect do
-          subject.call(name: "admin.books.add", slice: "main")
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
     end
   end
 end
