+++ spec/unit/hanami/cli/commands/app/generate/part_spec.rb
@@ -3,10 +3,9 @@
 require "hanami"
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Part, :app do
-  subject { described_class.new(fs: fs, out: out, err: err) }
+  subject { described_class.new(fs: fs, out: out) }
 
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:inflector) { Dry::Inflector.new }
   let(:app) { Hanami.app.namespace }
@@ -15,8 +14,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Part, :app do
   def output
     out.rewind && out.read.chomp
   end
-
-  def error_output = err.string.chomp
 
   context "generating for app" do
     context "without base part" do
@@ -57,25 +54,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Part, :app do
 
           expect(fs.read("app/views/parts/user.rb")).to eq(part)
           expect(output).to include("Created app/views/parts/user.rb")
-        end
-      end
-
-      context "with existing file" do
-        let(:file_path) { "app/views/parts/user.rb" }
-
-        before do
-          within_application_directory do
-            fs.write(file_path, "existing content")
-          end
-        end
-
-        it "exits with error message" do
-          expect do
-            within_application_directory { subject.call(name: "user") }
-          end.to raise_error SystemExit do |exception|
-            expect(exception.status).to eq 1
-            expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-          end
         end
       end
     end
@@ -225,26 +203,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Part, :app do
 
           # This is still printed because the fs.write above still prints
           # expect(output).to_not include("Created slices/main/views/part.rb")
-        end
-      end
-    end
-
-    context "with existing file" do
-      let(:file_path) { "slices/main/views/parts/user.rb" }
-
-      before do
-        within_application_directory do
-          fs.mkdir("slices/main")
-          fs.write(file_path, "existing content")
-        end
-      end
-
-      it "exits with error message" do
-        expect do
-          within_application_directory { subject.call(name: "user", slice: "main") }
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
         end
       end
     end
