+++ spec/unit/hanami/cli/commands/app/generate/repo_spec.rb
@@ -1,18 +1,15 @@
 # frozen_string_literal: true
 
 RSpec.describe Hanami::CLI::Commands::App::Generate::Repo, :app do
-  subject { described_class.new(fs: fs, out: out, err: err) }
+  subject { described_class.new(fs: fs, out: out) }
 
   let(:out) { StringIO.new }
-  let(:err) { StringIO.new }
   let(:fs) { Hanami::CLI::Files.new(memory: true, out: out) }
   let(:app) { Hanami.app.namespace }
 
   def output
     out.string
   end
-
-  def error_output = err.string.chomp
 
   context "generating for app" do
     describe "without namespace" do
@@ -94,23 +91,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Repo, :app do
       expect(fs.read("app/repos/books/published/hardcover_repo.rb")).to eq(repo_file)
       expect(output).to include("Created app/repos/books/published/hardcover_repo.rb")
     end
-
-    context "with existing file" do
-      let(:file_path) { "app/repos/book_repo.rb" }
-
-      before do
-        fs.write(file_path, "existing content")
-      end
-
-      it "exits with error message" do
-        expect do
-          subject.call(name: "books")
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
-    end
   end
 
   context "generating for a slice" do
@@ -152,23 +132,6 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Repo, :app do
 
       expect(fs.read("slices/main/repos/book/draft_repo.rb")).to eq(repo_file)
       expect(output).to include("Created slices/main/repos/book/draft_repo.rb")
-    end
-
-    context "with existing file" do
-      let(:file_path) { "slices/main/repos/book_repo.rb" }
-
-      before do
-        fs.write(file_path, "existing content")
-      end
-
-      it "exits with error message" do
-        expect do
-          subject.call(name: "books", slice: "main")
-        end.to raise_error SystemExit do |exception|
-          expect(exception.status).to eq 1
-          expect(error_output).to eq Hanami::CLI::FileAlreadyExistsError::ERROR_MESSAGE % {file_path:}
-        end
-      end
     end
   end
 end
